backend/requirements.txt
Flask
pandas
requests
transformers
google-cloud-videointelligence
